Add HEALTHCHECK directive
This commit adds HEALTHCHECK instructions to the Dockerfiles.

The actual healthcheck logic is in the `healthcheck.sh` file, which is
copied into the image via a `COPY` instruction in the Dockerfiles. The
reason for the separate script vs. inline in the Dockerfile is that
AFAIK, environment replacement isn't supported by the `HEALTCHECK`
instruction.